gasp: split into main and helper script
We've extracted the key part of the functionality to its own function,
`hip` (Hunks In Patch). `hip` is the middleman between a git diff and a
bunch of files, each with a single hunk within them. This makes `gaap`
(`Git Add A Patch`) quite light.

The plan is now to add alternatives to `gaap`, in `guap` and `gmap`
- Git Unstage A Patch and Git Murder A Patch respectively. Names not
final.