satod: don't show `gfp` output in reverse
Previously, we would make a reverse patch, and THEN apply it. Now, we
make a patch, and reverse-apply it. This makes the `gfp` output more
sane, since it actually shows the patch that you're unapplying, not the
reverse version of it.

While we're here, we add some comments to each part to benefit those who
come here later.