proof of concept
A more interactive mode, where you select a file, and then the patches
within it. I want to next try moving this entire functionality into FZF
binds - so you can press <Enter> to apply all the changes to a file,
or <Space> to choose a specific change. Hopefully there's a way to make
it less FZF-hellish with strings to execute - maybe I can write the
internal stuff as separate commands to get around that. The real worry I
have is removing a file after we do stuff with it. Anyways, lots to do,
lots to see!