fmbl: remove
This only existed as its own command when ghp, gfp, and gkp all shared a
lot of code. It didn't make sense when I extracted the logic into satod,
and it definitely doesn't make sense now that we removed the commands
wrapping satod. I also cleaned up the args a lot - removed the old binds
setup for a more modern version, and removed --inline-info.